refactor: update proto location
